BAH-913 | Sowmika | Reset appointment status (#12)
* Sowmika, Saikumar | [Reset Appointment Status] | Allow status change from any status to scheduled

* Sowmika, Saikumar | [Reset Appointment Status] | [Reformat] commit

* Sowmika, Saikumar | [Reset Appointment Status] | Authenticate user to reset status to scheduled

* Sowmika, Saikumar | [Reset Appointment Status] | [Refactor and Reformat] commit

* Sowmika, Saikumar | [Reset Apointment Status] | IT for reset appointment status to scheduled from any status

* Saikumar, Sowmika | [Reset Appointment status] | Add IT to check privilege to reset appointment status to scheduled

* Saikumar, Sowmika | [Reset Appointment status] | Migration script to add Reset Appointment Status privilege

* Saikumar, Sowmika | [Reset appointment status] | [Refactor] Status validator to support resest functionality

* Sowmika, Saikumar | [Reset Appointment Status] |  [Refactor] commit

* Sowmika, Saikumar | [Reset Appointment Status] | Add new endpoint status-change in AppointmentsController (Same as chanegStatus in AppointmentController)

* Sowmika | [Reset Appointment Status] | Changes from PR review

* Sowmika | [Reset Appointment Status] | Changes from PR review

# Conflicts:
#	api/src/main/java/org/openmrs/module/appointments/service/impl/AppointmentsServiceImpl.java
#	api/src/main/resources/liquibase.xml
#	api/src/test/java/org/openmrs/module/appointments/service/impl/AppointmentsServiceImplTest.java
#	omod/src/main/java/org/openmrs/module/appointments/web/controller/AppointmentController.java
#	omod/src/test/java/org/openmrs/module/appointments/web/controller/AppointmentControllerIT.java
#	omod/src/test/java/org/openmrs/module/appointments/web/controller/AppointmentControllerTest.java
