+++ api/src/main/resources/liquibase.xml
@@ -539,5 +539,17 @@
 
     </changeSet>
 
+    <changeSet id="201808241255" author="Saikumar, Sowmika">
+        <preConditions onFail="MARK_RAN">
+            <sqlCheck expectedResult="0">
+                SELECT COUNT(*) FROM privilege WHERE privilege = 'Reset Appointment Status';
+            </sqlCheck>
+        </preConditions>
+        <comment>Add Reset Appointment Status privilege to reset appointments from any status to scheduled.</comment>
+        <insert tableName="privilege">
+            <column name="privilege" value="Reset Appointment Status"/>
+            <column name="description" value="With this privilege, users can reset appointments from any status to scheduled."/>
+            <column name="uuid" valueComputed="UUID()"/>
+        </insert>
+    </changeSet>
 </databaseChangeLog>
-
