+++ omod/src/main/java/org/openmrs/module/appointments/web/controller/AppointmentController.java
@@ -122,25 +122,6 @@ public class AppointmentController {
         return appointmentsSummaryList;
     }
 
-    @RequestMapping(method = RequestMethod.POST, value="/{appointmentUuid}/changeStatus")
-    @ResponseBody
-    public ResponseEntity<Object> transitionAppointment(@PathVariable("appointmentUuid")String appointmentUuid, @RequestBody Map<String, String> statusDetails) throws ParseException {
-        try {
-            String toStatus = statusDetails.get("toStatus");
-            Date onDate = DateUtil.convertToLocalDateFromUTC(statusDetails.get("onDate"));
-            Appointment appointment = appointmentsService.getAppointmentByUuid(appointmentUuid);
-            if (appointment != null) {
-                appointmentsService.changeStatus(appointment, toStatus, onDate);
-                return new ResponseEntity<>(appointmentMapper.constructResponse(appointment), HttpStatus.OK);
-            } else {
-                throw new RuntimeException("Appointment does not exist");
-            }
-        } catch (RuntimeException e) {
-            log.error("Runtime error while trying to update appointment status", e);
-            return new ResponseEntity<>(RestUtil.wrapErrorResponse(e, e.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
     @RequestMapping(method = RequestMethod.POST, value="undoStatusChange/{appointmentUuid}")
     @ResponseBody
     public ResponseEntity<Object> undoStatusChange(@PathVariable("appointmentUuid")String appointmentUuid) throws ParseException {
@@ -202,4 +183,5 @@ public class AppointmentController {
             return new ResponseEntity<>(RestUtil.wrapErrorResponse(e, e.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
+
 }
