+++ omod/src/test/java/org/openmrs/module/appointments/web/controller/AppointmentControllerIT.java
@@ -72,9 +72,9 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
                 "\"endDateTime\": \"2017-07-20\",  " +
                 "\"appointmentKind\": \"WalkIn\", " +
                 "\"providers\": [ {" +
-                "\"uuid\":\"2d15071d-439d-44e8-9825-aa8e1a30d2a2\"," +
-                "\"comments\":\"available\"," +
-                "\"response\":\"ACCEPTED\"" +
+                    "\"uuid\":\"2d15071d-439d-44e8-9825-aa8e1a30d2a2\"," +
+                    "\"comments\":\"available\"," +
+                    "\"response\":\"ACCEPTED\"" +
                 "} ] }";
 
         MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment", content));
@@ -88,8 +88,7 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
         String serviceTypeUuid = "678906e5-9fbb-4f20-866b-0ece24564578";
         MockHttpServletRequest getRequest = newGetRequest(requestURI, new Parameter("appointmentServiceTypeUuid", serviceTypeUuid));
         List<AppointmentDefaultResponse> asResponse = deserialize(handle(getRequest),
-                new TypeReference<List<AppointmentDefaultResponse>>() {
-                });
+            new TypeReference<List<AppointmentDefaultResponse>>() {});
 
         assertEquals(5, asResponse.size());
     }
@@ -110,7 +109,7 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
         assertEquals("c36006e5-9fbb-4f20-866b-0ece245615a6", appointmentsSummary.getAppointmentService().getUuid());
         assertEquals(1, appointmentsSummary.getAppointmentCountMap().size());
         Map<String, DailyAppointmentServiceSummary> appointmentCountMap = appointmentsSummary.getAppointmentCountMap();
-        Map appointmentCount = (Map) appointmentCountMap.get("2108-08-15");
+        Map appointmentCount = (Map)appointmentCountMap.get("2108-08-15");
         assertNotNull(appointmentCount);
         assertEquals(4, appointmentCount.get("allAppointmentsCount"));
         assertEquals(1, appointmentCount.get("missedAppointmentsCount"));
@@ -154,57 +153,6 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
     }
 
     @Test
-    public void should_changeAppointmentStatusWithDate() throws Exception {
-        String onDate = "2108-08-22T10:30:00.0Z";
-        String content = "{ \"toStatus\": \"CheckedIn\", \"onDate\":\"" + onDate + "\"}";
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/c36006e5-9fbb-4f20-866b-0ece245615a7/changeStatus", content));
-        assertNotNull(response);
-        Appointment appointmentByUuid = appointmentsService.getAppointmentByUuid("c36006e5-9fbb-4f20-866b-0ece245615a7");
-        assertEquals(200, response.getStatus());
-        assertNotNull(appointmentByUuid);
-        List<AppointmentAudit> historyForAppointment = appointmentAuditDao
-                .getAppointmentHistoryForAppointment(appointmentByUuid);
-        assertEquals(1, historyForAppointment.size());
-        assertNotNull(historyForAppointment.get(0).getId());
-        assertNotNull(historyForAppointment.get(0).getNotes());
-        assertEquals(DateUtil.convertToLocalDateFromUTC(onDate).toInstant().toString(), historyForAppointment.get(0).getNotes());
-    }
-
-    @Test
-    public void should_changeAppointmentStatusWithoutDate() throws Exception {
-        String content = "{ \"toStatus\": \"CheckedIn\"}";
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/c36006e5-9fbb-4f20-866b-0ece245615a7/changeStatus", content));
-        assertNotNull(response);
-        Appointment appointmentByUuid = appointmentsService.getAppointmentByUuid("c36006e5-9fbb-4f20-866b-0ece245615a7");
-        assertEquals(200, response.getStatus());
-        assertNotNull(appointmentByUuid);
-        List<AppointmentAudit> historyForAppointment = appointmentAuditDao
-                .getAppointmentHistoryForAppointment(appointmentByUuid);
-        assertEquals(1, historyForAppointment.size());
-        assertNotNull(historyForAppointment.get(0).getDateCreated());
-        assertNotNull(historyForAppointment.get(0).getCreator());
-        assertEquals(appointmentByUuid, historyForAppointment.get(0).getAppointment());
-        assertEquals(appointmentByUuid.getStatus(), historyForAppointment.get(0).getStatus());
-        assertNull(historyForAppointment.get(0).getNotes());
-    }
-
-    @Test
-    public void should_throwExceptionForInvalidStatusChange() throws Exception {
-        String content = "{ \"toStatus\": \"Scheduled\"}";
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/c36006e5-9fbb-4f20-866b-0ece245615a7/changeStatus", content));
-        assertNotNull(response);
-        assertEquals(400, response.getStatus());
-    }
-
-    @Test
-    public void should_throwExceptionForInvalidAppointment() throws Exception {
-        String content = "{ \"toStatus\": \"Scheduled\"}";
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/c36006e5-9fbb-4f20-866b-0ece245615a8/changeStatus", content));
-        assertNotNull(response);
-        assertEquals(400, response.getStatus());
-    }
-
-    @Test
     public void shouldCreateAuditEventsWhenDetailsChangesOnEditAppointment() throws Exception {
         String content = "{ \"uuid\": \"c36006e5-9fbb-4f20-866b-0ece245615a7\", " +
                 "\"appointmentNumber\": \"1\",  " +
@@ -231,58 +179,12 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
 
     @Test
     public void shouldThrowExceptionForInvalidAppointmentOnUndoStatus() throws Exception {
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/" + "randomUuid", "{}"));
+        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/"+ "randomUuid", "{}"));
         assertNotNull(response);
         assertEquals(400, response.getStatus());
         assertTrue(response.getContentAsString().contains("Appointment does not exist"));
     }
 
-    @Test
-    public void shouldUndoCheckedInAppointment() throws Exception {
-        String content = "{ \"providerUuid\": \"823fdcd7-3f10-11e4-adec-0800271c1b75\", " +
-                "\"appointmentNumber\": \"3\",  " +
-                "\"patientUuid\": \"2c33920f-7aa6-48d6-998a-60412d8ff7d5\", " +
-                "\"serviceUuid\": \"c36006d4-9fbb-4f20-866b-0ece245615c1\", " +
-                "\"startDateTime\": \"2017-07-20\", " +
-                "\"endDateTime\": \"2017-07-20\",  " +
-                "\"appointmentKind\": \"WalkIn\"}";
-
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment", content));
-        assertNotNull(response);
-        assertEquals(200, response.getStatus());
-        AppointmentDefaultResponse appointmentDefaultResponse = deserialize(response, new TypeReference<AppointmentDefaultResponse>() {
-        });
-        String appointmentUuid = appointmentDefaultResponse.getUuid();
-
-        content = "{ \"toStatus\": \"CheckedIn\"}";
-        response = handle(newPostRequest("/rest/v1/appointment/" + appointmentUuid + "/changeStatus", content));
-        assertEquals(200, response.getStatus());
-        content = "{}";
-        response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/" + appointmentUuid, content));
-        assertNotNull(response);
-        assertEquals(200, response.getStatus());
-        Appointment appointment = appointmentsService.getAppointmentByUuid(appointmentUuid);
-        assertNotNull(appointment);
-        assertEquals("Scheduled", appointment.getStatus().toString());
-    }
-
-    @Test
-    public void shouldUndoMissedAppointment() throws Exception {
-        String appointmentUuid = "c36006e5-9fbb-4f20-866b-0ece245615a7";
-        String content = "{ \"toStatus\": \"CheckedIn\"}";
-        MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment/" + appointmentUuid + "/changeStatus", content));
-        assertEquals(200, response.getStatus());
-        content = "{ \"toStatus\": \"Missed\"}";
-        response = handle(newPostRequest("/rest/v1/appointment/" + appointmentUuid + "/changeStatus", content));
-        assertEquals(200, response.getStatus());
-        content = "{}";
-        response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/" + appointmentUuid, content));
-        assertNotNull(response);
-        assertEquals(200, response.getStatus());
-        Appointment appointment = appointmentsService.getAppointmentByUuid(appointmentUuid);
-        assertNotNull(appointment);
-        assertEquals("CheckedIn", appointment.getStatus().toString());
-    }
 
     @Test
     public void shouldThrowExceptionWhenThereIsNoPriorActionToUndo() throws Exception {
@@ -296,11 +198,10 @@ public class AppointmentControllerIT extends BaseIntegrationTest {
         MockHttpServletResponse response = handle(newPostRequest("/rest/v1/appointment", content));
         assertNotNull(response);
         assertEquals(200, response.getStatus());
-        AppointmentDefaultResponse appointmentDefaultResponse = deserialize(response, new TypeReference<AppointmentDefaultResponse>() {
-        });
+        AppointmentDefaultResponse appointmentDefaultResponse = deserialize(response, new TypeReference<AppointmentDefaultResponse>() {});
         String appointmentUuid = appointmentDefaultResponse.getUuid();
         content = "{}";
-        response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/" + appointmentUuid, content));
+        response = handle(newPostRequest("/rest/v1/appointment/undoStatusChange/"+ appointmentUuid, content));
         assertNotNull(response);
         assertEquals(400, response.getStatus());
         assertTrue(response.getContentAsString().contains("No status change actions to undo"));
